refactoring & handling edge cases with hyperparam mutations
